+++ test/writer_test.cpp
@@ -455,20 +455,17 @@ TEST_CASE("Writer Pages", "[Writer]")
         REQUIRE(!writer.FindNode(VoxelKey::RootKey()).IsValid());
         REQUIRE(!writer.FindNode(VoxelKey::InvalidKey()).IsValid());
         REQUIRE(!writer.FindNode(VoxelKey(5, 4, 3, 2)).IsValid());
-
-        REQUIRE_NOTHROW(writer.GetRootPage());
-        Page root_page = writer.GetRootPage();
-        REQUIRE(root_page.IsValid());
-        REQUIRE(root_page.IsPage());
-        REQUIRE(root_page.loaded == true);
-
-        REQUIRE_THROWS(writer.AddSubPage(root_page, VoxelKey::InvalidKey()));
+        auto header = *writer.CopcConfig()->LasHeader();
+        las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
+        points.AddPoint(points.CreatePoint());
+        // Add a node with root key as page
+        writer.AddNode(VoxelKey(1, 1, 1, 1), points, VoxelKey::RootKey());
 
         writer.Close();
 
         Reader reader(&out_stream);
         REQUIRE(reader.CopcConfig().CopcInfo().root_hier_offset > 0);
-        REQUIRE(reader.CopcConfig().CopcInfo().root_hier_size == 0);
+        REQUIRE(reader.CopcConfig().CopcInfo().root_hier_size == 32);
         REQUIRE(!reader.FindNode(VoxelKey::InvalidKey()).IsValid());
     }
 
@@ -478,22 +475,55 @@ TEST_CASE("Writer Pages", "[Writer]")
 
         Writer writer(out_stream, {6});
 
-        Page root_page = writer.GetRootPage();
+        auto header = *writer.CopcConfig()->LasHeader();
+        las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
+        points.AddPoint(points.CreatePoint());
 
-        auto sub_page = writer.AddSubPage(root_page, VoxelKey(1, 1, 1, 1));
-        REQUIRE(sub_page.IsPage());
-        REQUIRE(sub_page.IsValid());
-        REQUIRE(sub_page.loaded == true);
-
-        REQUIRE_THROWS(writer.AddSubPage(sub_page, VoxelKey(1, 1, 1, 0)));
-        REQUIRE_THROWS(writer.AddSubPage(sub_page, VoxelKey(2, 4, 5, 0)));
+        writer.AddNode(VoxelKey(1, 1, 1, 1), points, VoxelKey(1, 1, 1, 1));
+        writer.AddNode(VoxelKey(3, 4, 4, 4), points, VoxelKey(2, 2, 2, 2));
+        writer.AddNode(VoxelKey(2, 0, 2, 2), points, VoxelKey(1, 0, 1, 1));
 
         writer.Close();
 
         Reader reader(&out_stream);
         REQUIRE(reader.CopcConfig().CopcInfo().root_hier_offset > 0);
-        REQUIRE(reader.CopcConfig().CopcInfo().root_hier_size == 32);
+        REQUIRE(reader.CopcConfig().CopcInfo().root_hier_size == 64); // size of two sub pages of the root page
         REQUIRE(!reader.FindNode(VoxelKey::InvalidKey()).IsValid());
+
+        auto page_keys = reader.GetPageList();
+        REQUIRE(page_keys.size() == 4);
+        REQUIRE(std::find(page_keys.begin(), page_keys.end(), VoxelKey(1, 1, 1, 1)) != page_keys.end());
+        REQUIRE(std::find(page_keys.begin(), page_keys.end(), VoxelKey(2, 2, 2, 2)) != page_keys.end());
+        REQUIRE(std::find(page_keys.begin(), page_keys.end(), VoxelKey(1, 0, 1, 1)) != page_keys.end());
+    }
+
+    SECTION("Change Node Page")
+    {
+        stringstream out_stream;
+
+        Writer writer(out_stream, {6});
+
+        auto header = *writer.CopcConfig()->LasHeader();
+        las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
+        points.AddPoint(points.CreatePoint());
+
+        writer.AddNode(VoxelKey(3, 4, 4, 4), points, VoxelKey(2, 2, 2, 2));
+
+        writer.ChangeNodePage(VoxelKey(3, 4, 4, 4), VoxelKey(1, 1, 1, 1));
+
+        // Check for validity
+        REQUIRE_THROWS(writer.ChangeNodePage(VoxelKey(3, 4, 4, 3), VoxelKey(1, 0, 0, 0)));  // Node doesn't exist
+        REQUIRE_THROWS(writer.ChangeNodePage(VoxelKey(3, 4, 4, -1), VoxelKey(1, 0, 0, 0))); // Invalid Node
+        REQUIRE_THROWS(writer.ChangeNodePage(VoxelKey(3, 4, 4, 4), VoxelKey(1, 0, 0, -1))); // Invalid New Page
+
+        // Check for parental link
+        REQUIRE_THROWS(writer.ChangeNodePage(VoxelKey(3, 4, 4, 4), VoxelKey(1, 0, 0, 0)));
+
+        writer.Close();
+
+        Reader reader(&out_stream);
+        auto node = reader.FindNode(VoxelKey(3, 4, 4, 4));
+        REQUIRE(node.page_key == VoxelKey(1, 1, 1, 1));
     }
 }
 
@@ -590,12 +620,10 @@ TEST_CASE("Writer Copy", "[Writer]")
 
         Writer writer(out_stream, cfg);
 
-        Page root_page = writer.GetRootPage();
-
         for (const auto &node : reader.GetAllNodes())
         {
             // only write/compare compressed data or otherwise tests take too long
-            writer.AddNodeCompressed(root_page, node.key, reader.GetPointDataCompressed(node), node.point_count);
+            writer.AddNodeCompressed(node.key, reader.GetPointDataCompressed(node), node.point_count);
         }
 
         writer.Close();
@@ -632,7 +660,6 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
 
@@ -642,7 +669,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         point->Z(5);
 
         points.AddPoint(point);
-        writer.AddNode(root_page, {1, 1, 1, 1}, points);
+        writer.AddNode({1, 1, 1, 1}, points);
         writer.Close();
 
         FileReader reader(file_path);
@@ -655,7 +682,6 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
 
@@ -665,7 +691,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         point->Z(5.1);
 
         points.AddPoint(point);
-        writer.AddNode(root_page, {2, 3, 3, 3}, points);
+        writer.AddNode({2, 3, 3, 3}, points);
         writer.Close();
 
         FileReader reader(file_path);
@@ -678,7 +704,6 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
 
@@ -688,7 +713,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         point->Z(5.1);
 
         points.AddPoint(point);
-        writer.AddNode(root_page, {1, 1, 1, 1}, points);
+        writer.AddNode({1, 1, 1, 1}, points);
         writer.Close();
 
         FileReader reader(file_path);
@@ -701,7 +726,6 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
 
@@ -711,7 +735,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         point->Z(0.1);
 
         points.AddPoint(point);
-        writer.AddNode(root_page, {1, 0, 0, 0}, points);
+        writer.AddNode({1, 0, 0, 0}, points);
         writer.Close();
 
         FileReader reader(file_path);